Move OAuth secret management into pkg/auth/oauth
Previously pkg/oauth/ contained only OAuth client secret management
functions (ProcessOAuthClientSecret, StoreSecretInManager), which
was confusing because it shared a similar name with pkg/auth/oauth/
which contains the actual OAuth/OIDC protocol implementation.

This change:
- Moves pkg/oauth/client_secret.go → pkg/auth/oauth/secrets.go
- Updates 2 call sites: cmd/thv/app/run_flags.go, pkg/runner/config.go
- Removes pkg/oauth/ directory entirely

Benefits:
- Clearer package structure (all OAuth code under pkg/auth/oauth/)
- No more confusion between pkg/oauth and pkg/auth/oauth
- Better separation: secrets.go for secret management, flow.go for OAuth protocol
